Renamed file plotlyQueryHooks.ts to queryHooks.ts
