+++ frontend/src/modules/SimulationTimeSeries/queryHooks.tsx
